Add fields in OdrefreshReported for compilation times in milliseconds
Currently we are reporting the compilation times for the boot
classpath for primary/secondary architecture and system server
in seconds, but since the values we have observed so far are
4~8 seconds for the former and 8~16 for the latter, we might
benefit from having a finer granularity. Hence, this commit
is producing the compilation time in milliseconds (alongside
the previous values in seconds, for a transition period).

Bug: 243009337
Test: atest ArtGtestTargetChroot
Change-Id: Ided03b350fe9f48f06a2a22d622b3e731d660cf4