Modify ApexInfoGathered atom.
The existing ApexInfoGathered atom collects the SHA256 digest of
a file. This becomes very computationally expensive with more binaries.

We therefore adjust this atom to make use of a more flexible digest
computation scheme that can still accommodate SHA256 in the future.

We also introduce a new field to capture the digest of the signer
cert of the APEX.

Bug: 245944675

Test: Manual. A related change is also submitted to fix the call
to logging so as to not affect compilation.

Change-Id: Ide6a48553d55ae960a30f4bd45fd2d2595dcb019