Add new atom to log wakeup attribution
A new atom for each reason attributed with a wakeup. Whenever a kernel
wakeup happens, it may be attributed to a few top level reasons. Each of
them can be described with this atom.

The atom is split per reason to avoid complications with nesting
messages making it easier to filter and aggregate.

Test: statsd_testdrive 588

Bug: 195684213
Bug: 265737334
Change-Id: I1c06d01811a9fc9827a2081c6004ae4ecca0f0a6