Atom: move ADPFSystemComponentReported to pull atom
- Move ADPFSystemComponentReported to pull atom
- Rename surfaceflinger_enabled to surfaceflinger_cpu_hint_enabled

Given the low reboot rate of mobile device, the previous design
of the pushed atom that push the data per device boot will get us
low amount of data. As we want to measure the usage of these two
components on a daily/weekly basis, it is more reasonable to change
it to a pull atom that pulls daily.

Bug: 261639793
Bug: 260574648
Test: TBD
Change-Id: Ic130ce9786d737cb596cdc226d9ca80d38666c12