Rename placeholder_atoms.proto to placeholder_extension_atoms.proto
In the new folder structure, extension and non-extension atoms need to be kept in separate files
because extension atoms depend on atoms.proto and non-extension atoms are a dependency for
atoms.proto. Combining extension and non-extension atoms in the same file will cause a dependency
cycle. So we introduce a different filename format for extension atoms.

Bug: 256259429
Test: TH
Change-Id: I52ba9ea222259bfdcc556310f3cc2db769467a9a