Adding V3 Metrics with Further Bits
This aims to split the atom but does not yet add strings, as our current
decided actions are to add strings in V4. This does, however, add all
the other bits that we decided on over the metric overview doc, such as
bits of 'arrival'. It also splits the atom and uses sequence ids and
session ids to allow coalescing it in the backend. More will be added
over time.

Bug: 270403549
Test: Will be chained in the future
Change-Id: I50b60abe9014191d3aa3f502485e190b52a55ef2