Changing SessionID Size and Setting Metrics
To be consistent across flows, we plan to switch to a session Id generated by the
Binder. Also, we need the new metric atoms to be visible to the
frameowrk, so a few bits needed to be flipped in the atoms file. Doing
both here. An enum was also redundantly repeated, so that was also
removed in this change, as it was affecting how the proto built.

Bug: 271135048
Bug: 270403549
Test: Builds for now
Change-Id: I2f2d55a5cd53282f5870fce6e83cef62c6164c09