Add Cronet's atoms to AOSP
Now that Cronet is shipped as part of the Tethering mainline module,
it makes more sense to move its atoms IDs in the platform range.

The atoms here have been copied 1:1 from
https://source.corp.google.com/search?q=f:cronet_atoms.proto.
If copied downstream as is, these would clash with the pre-existing
ones in the app range (since the extension and proto field will have
the same name).
To solve this issue, we will deprecate the existing atoms in the app range
and have Cronet always log using the IDs of this extension.

Test: I've checked statsd output for a local build with logging enabled
Bug: 284972285

Change-Id: I4306f311143151fd0f26176f6a837229a7989416