Input proto containing Keyboard atoms
Created input_extension_atoms.proto to contain atoms related to input.
Added the KeyboardConfigured atom to log keyboard layouts configured
when a device is connected, as well as whether it is a first time
configuration.
Added the KeyboardSystemsEventReported atom to log shortcut usage on the
physical keyboard.
Created enums.proto for input to contain enums related to input atoms.

Bug: 280423320, 280603577
Test: Build
Change-Id: I07d6ec04e714f7c34f98c4abfe98d983da969d1a