Add Keyboard Enums
Add the LayoutSelectionCriteriaKeyboard and KeyboardLayoutType enums.
Deprecated the KeyboardLayoutSelectionCriteria enum since Enums should
include a default UNSPECIFIED value as the first value in the
declaration. 0 being mapped to a value other than UNSPECIFIED was causing
unexpected behaviour (such as the field not showing up in the atom) when logged the atom.

Bug: 280603577
Test: Build

Change-Id: I03440404c58c33820f465168eeca5bb9246e0236