Add new methods and stages to SandboxApiCalled
New methods are added to match new APIs called from SDK.
New stages are also added to collect total call latency. There are 2
stages as we need a way to differentiate between cases when sandbox was
loaded during the call and when it wasn't (applies only to loadSdk()).
Eldar doc: eldar/227133964

Bug: 288864263, Bug: 294216354
Test: manual
Change-Id: I9f4ea9d2749eaad59910f1ac04789d866d8ad6fa