Sync up measurement atoms to match internals
Bug: 297120274
The measurement atom values are not in sync with the value from the
internal repository. This value was added to debug an issue happening
when a caller app has the cloned apps feature. Measurement code threw an
exception when verifying the foreground checks as the caller app was a
cloned app.

Test: mma
Change-Id: I13b01546d2ae359e9cf7537afc1380ea68d67b5c