[Cherry Pick] Refactor Cronet's atom
See b/302452793 for MDR and approvals, This is just a cherry-pick of the internal version. Cronet is developed in AOSP and we need to cherry-pick the atom definition to AOSP in order for it to work.

--------------------------
* Added a new pushed atom [CronetInitialized] that will be logged everytime Cronet is
initialized for the very first time. This is necessary to unblock the development of Cronet Experiment framework
and gain more insights.
* Refactor the older atoms and move some fields to the new atoms. The atoms are still backward compatible.
* See cl/568191940 for more information about the metrics and changes.
* See b/302452793 for the approval of the new metrics along with an Android-metrics-eldar-document.

Bug: 303216384
Bug: 288088422
Bug: 304481024
Test: m libstats_atoms_proto
Merged-In: Ie19a233670a2d6daaaf5fa671696d81113e36f11
Change-Id: Ic79f6d3e72d6476c3bf0d0b4f2d6f87f8a75ca0e