Change drawable_type field from enum to int32.
The problem with the existing implementation is that it requires logging
the DrawableRendered atoms multiple times for each rendered layout. The
atoms need to be changed to log the aggregated data of multiple atoms.

By changing drawable_type from an enum to int32, we can now log the
existing of multiple drawable type in a single int value (bitset).

BUG: 309920856

Change-Id: I0360186c662f665204d0212ba3d8541630b2cf12