Creates TEX metrics for Accessibility FAB actions: dismiss & edit.
Includes the UID of the acted-upon shortcut features so that logs
include which features were modified.

These metrics allow the Accessibility team to see which features are
being edited to use another shortcut vs. which are being fully
dismissed, and to compare the rates of dismiss vs. edit (per feature)
to understand user preferences for the FAB for a given accessibility
feature.

Test: Observe logged metrics using 'statsd_testdrive 644'
Bug: 324968169
Change-Id: If979b586b3d3791b36eff5006d161b5d2a54317c