[Autofill Log] : Proto changes for logging views filling
Add logs for tracking failures and success on views filling.
Note that for the success cases, these are the upper bound. It's
calculated when the framework sends the autofill values to the view.
It's possible that the views reject autofill for some reason, and
doesn't autofill. Since autofill doesn't have any way to know that,
we log what we can.

Bug: 319872477
Test: m
Change-Id: I053d3ea3055df8d82dfb7b1b2bf345951fedc658