Create TeX metrics for Accessibility Quick Settings shortcut
Log the metric as a counter with uid where the uid indicates where the
user triggers the action. Currently it can be either Setting (who shares
the uid with android system) or SystemUI uid.

These metrics allow the Accessibility team to understand where the user
usually go to configure the quick setting shortcut.

Bug: 330572335
Test: Run `statsd_testdrive 644` and verify the output matches
Flag: N/A
Change-Id: Ib4672dd0d56dd21b3b9344a075b1b7b8fb359318