atoms: Add microxr atom
We are moving our first atom to the platform id range so now we are
adding them to the proto_logging library. Here we are adding the Android
counterpart with the MicroXRDeviceBootCompleteReported  atom definiton.

Flag: NONE moving an atom from the vendor range to the platform one
Bug: b/339691918
Tested: build. Also see Iedaf95ddf8619013b3eeb1467bb5edcb3978c7ba .
Change-Id: I1d0f00ea8674de9b8569029b5d136db9a928824f