Add ChangeOrigin enum and field to DNDStateChanged.
This mirrors the ZenModeConfig.ConfigChangeOrigin enum and indicates the source of a change to the overall zen system.

Eldar: https://eldar.corp.google.com/assessments/332763101

While the change to the proto definitions is unflaggable, these fields will only be populated when the flag android.app.modes_ui is true.

Flag: unflaggable
Bug: 316950757
Test: statsd_testdrive
Change-Id: I8f5b80c6523e8893824d89403bbf63c68343ead1