Adds TestUprobeStatsAtomReported
This atom is similar to TestAtomReported (intended only for testing,
not logged anywhere). We create a separate atom for
packages/modules/UprobeStats because that module will have its own tests
and we don't want to break each other's tests with unrelated atom
changes.

Bug: 296108553
Test: TH
Flag: NONE UprobeStats does not yet run on devices
Change-Id: I3256afe763c215e989bf6ba78e92175b99e81569