Add QuirkEvent for PD provisioned
Since other attempts at creating a provisioning mode metric are not moving forward, this just takes the much simpler approach of adding a quirk event instead. This is trivial to integrate in the current system, and will at least provide some metrics about whether the PD code is actually working.

Change-Id: I86568316cd979cd1e9e740713c2a98697fa197cb
Test: TH