Add a new atom to log AppOp access history for certain AppOps
We want to have a better visibility of AppOp access patterns to help
make informed decision of improving AppOp historical access data storage and elevate AppOp usage monitoring.

Create a new pushed atom to log AppOp accesses when noteOp, startOp or finishOp are called. Only a limited amount of AppOps will be logged, which includes AppOps that are currently in DiscreteRegistry(e.g. LOCATION, CAMERA, MICROPHONE) or are candidates to be included in DiscreteRegistry.

Flag: EXEMPT log only update
Bug: 365584286
Test: m
Change-Id: I51e22a5659ba149cff10a64def77e5db2887fc81