Add atom for Classification adjustments
Adds atom to store information about notification channel
reclassifications.
Stores information on whether the notification had already been posted
or was alerting when the reclassification occurs,
the new suggested type, and the notification's lifetime when the adjustment
is received, which tells us latency of the classification.

Bug: 375462619
Test: build
Flag: EXEMPT resource only update
Change-Id: I29dced409903576e0da29b25d292eb7a53d88ebe