Add atom for Notification Bundle Modification
To be used in System UI to record when a notification bundle is
dismissed. Records the bundle type and whether the bundle's contents
were all shown. Package and channel information is identical to
channel identifying information recorded in
NotificationChannelInteracted
to allow the records to be aligned.

Bug: 375462619
Test: mp droid
Flag: EXEMPT proto change
Change-Id: Ifd495631c5a3b2235d843874c5bf589672a35b87