Update CertificateTransparencyLogListUpdateFailed atom
This atom was previously added in ag/29792068. This change extracts the enums into a separate file (with OWNERS) and also adds a field for HTTP error codes, if applicable.

Note: this proto has not yet been used or logged anywhere, which is why we are okay to modify it.

Bug: 378626065
Flag: com.android.net.ct.flags.certificate_transparency_service
Test: builds

Change-Id: I1ef5b55b9a4dd0df2fa4f42f7783f67548923ea7