Update documentation of Keystore2 atoms.
I'm removing comments from several enums since the AIDL enums that
they mirror are the source of truth about the enum values' semantics.
Values should only be commented if they deviate from the AIDL spec.

The "Keystore2KeyCreationWithAuthInfo.user_auth_type" field doesn't
contain the expected information since the corresponding AIDL type is
a bitmask, not an enum. For now I'm updating the field's comment to
faithfully describe how this metrics enum is structured, but I'll fix
the field and the comment in this file in a separate commit.

Bug: 385175793
Test: n/a (comment updates)
Change-Id: I1990af3babf47dbf1073ab99cd72d02f7d69c8ed