Add enum for tethering with placeholder request
This enum audits the times tethering is started with a placeholder
request. Placeholder requests are used whenever we get a link layer
event to trigger tethering but we can't find a pending request for the
interface. This may happen due to race conditions such as stopTethering
removing the pending request right after startTethering but before the
link layer event, or IPv4 prefix conflict restarting tethering.

Bug: 216524590
Test: build
Change-Id: Ic730c331a3d0e4d8df7ac30689b65571782dec7f