Add ZramMmStatMmd and ZramBdStatMmd pulled atoms
mmd will report pulled atoms for zram stats; ZramMmStatMmd and
ZramBdStatMmd. These atoms are "Mmd" suffixed because we already have
ZramMmStat and ZramBdStat which are Pixel vendor specific pushed atoms.

The existing atoms are implemented as pushed atoms because pulled atoms
are not supported for vendor metrics.

This introduce the new metrics with different names and will deprecate
the Pixel vendor specific pushed atoms later.

Also the units of all fields of ZramMmStatMmd and ZramBdStatMmd are
KB while the raw values of kernel are in bytes or pages. The acutal
size in KB is preferred because some platform may use different page
size. Reporting metrics in bytes is unnecessary because data is
sampled and metrics are estimated. smaller metrics are cheaper to
process and store.

Bug: 392783204
Test: manual
Flag: android.mmd.flags.mmd_enabled
Change-Id: Ida14ad849cdef73eca53aedd47f4964705193148