Add HttpEngine preload count metric
This will show us the total number of times HttpEngine got preloaded on
different devices, the metric will be called from HttpEngine#Preload.
However, HttpEngine#Preload is being called from the Zygote only so
this effectively is being counted everytime the Zygote preloads
HttpEngine.

This is a normal westworld atom which means we'll get to see a breakdown
by SDK version to ensure that this is working across the fleet (S -> B).

Test: statsd_testdrive 528 and verified the metricID does show up + I
also did some manual printf to statds listener socket to ensure it's
actually being received.
Bug: 380349437
Change-Id: I2a82361e9a8a01acde01d96427f61448b63acb93

Change-Id: I80eb7e52a299497e7da87835bbce202f3e5c59df