[Metrics] Add Prepare API Boolean
This bit indicates that the API flow for this row of data utilizes the
prepare flow. While at the moment, the only API that does a prepare flow
is the 'get' API, this avoids having to rename the API_NAME to a brand
new API - instead, it allows users of queries to just add a quick new
clause to check for this boolean bit to know if it was a 'get' API built
using prepare, for example.

This avoids needing to modify RequestInfo to handle parsing an ApiName.

Bug: 393172214
Bug: 362995981
Test: Tested in Chained CL/StatsD Tested via DevTestApp
Change-Id: I18e72947fa2f02a535af89a7caf735d0114956a3
Flag: android.credentials.flags.fix_metric_duplication_emits