[Metrics] Add Chosen Credential String
This has been a much requested feature that will allow our queries to be
simplified tremendously. To be thorough, this was reviewed with PwG.
This ensures that we can, in both tracks, get the selected credential at
the end of the flow, completing the funnel.

Bug: 364444783
Test: Build, stats_d, todo chain
Flag: android.credentials.flags.fix_metric_duplication_emits
Change-Id: I3a279370d5e4da450754ea44b56295b8c9fd8ed4