Add list of denied adjustments to package preferences.
If an adjustment key (for example, KEY_TYPE or KEY_SUMMARIZATION) is included in this list it means the user has excluded this app for that kind of adjustment.

Also deprecates the field for bundle types allowed by package, as we won't launch with this fine-grained permission.

Bug: 409751981
Bug: 413434963
Test: atom tester
Flag: android.app.notification_classification_ui
Change-Id: If5d3e6ecac444234ca1db357d35ff0c4b2dd82de