[ADI][metrics] combine two constants into one
There is no need to have two separate constants just because the
correspondnig UI interfaces are slightly different. From the logging
perspective, the UI can be inferred from the verification result and the
verification policy, if needed. The purpose of this field is to log the
user response, for which we just need to know if the user has chosen to
abort, or retry, or install_anyway.

FLAG: android.content.pm.verification_service
BUG: 422400021
FIXES: 422400021
Test: to be added

Change-Id: I0cc4d132d85570c7c1259fc7e9ab5c6e005e1ef7