Change nested repeated field in logging proto
Replace the nested repeated proto field with two lists of enums. StatsD
does not have a graceful support for nested proto messages because
protos and enums defined in frameworks/proto_logging are not visible to
no test code.

Acknowledging that this is against practices of logging protos, because
of backwards compalibility, however this proto is not used anywhere yet,
the decision for this change was made while trying to use it for the
first time.

Bug: 400696164
Test: TH
Flag: EXEMPT statsd atoms/metrics changes
Change-Id: Ieb63e19486591fe42cff2ab67141ac711205f7c5