Add CallException errors to enums.proto
Adds new error enums to the ErrorEnum message to correspond with the
error codes defined in android.telecom.CallException. These exceptions
are used by the transactional call APIs.

This allows for more granular logging of failures related to
transactional calls, which helps in monitoring the stability and
performance of these new APIs.

Flag: NONE metrics
Fixes: 424246038
Test: build
Change-Id: Idb7e96e67e64439b83bd27ff34de72d962c45643