[USB Telemetry] Adding atom and enums for USB Telemetry
We are adding a new atom for logging Userspace AOA state in
USB. This will help us analyse what proportion of the devices are
using the feature. It will also help us analyse the reasons of
failure if any.

Bug: 444618795
Test: Manually tested by flashing the build and verifying that the atom
is logged correctly using `statsd_testdrive` and `adb logcat`.
Flag: EXEMPT METRICS

Change-Id: I005adba01cfe30952a23a6a6852fcbda83cf05a1