[metrics] updates PackageInstallerSessionReported
Adds a new long to PackageInstallerSessionReported, called
fs_verity_enabled_apks_size_bytes, which tracks the total size of
packages for which fs verity was set up.
This will be used to enabled A/B installation performance testing, so we
can compare the installation times across packages based on the status
of fs-verify. This metric will be used as part of a larger pack of
metric to evaluate the overall performance of fs-verity enabled APKs.

Flag: EXEMPT METRICS
Bug: 451741124
Test: `statsd_testdrive 1101`
Change-Id: I4eb5b03e3e3c44337d2d5612380fe885931763aa