[metrics] adds package_name to InitAppScanReported
Updates the atom for InitAppScanReported to include the newly added
field, pacakge_name. package_name records the name of the system app
being scanned during boot. This is needed to break down the metric by
app name, to monitor various dimensions, including the signature version
based on the app. Knowing the signature version used for verification of
a given system app in the fleet is an important data point for avoiding
performance regressions as we rollout various aspects of Full Stack
Integrity (FSI). This metric is only reported for updated system apps.

Bug: 462112193
Flag: EXEMPT METRICS
Test: m
Change-Id: Id0527f86f2cbdc655a6da00b01cb7209ba62ec72