Rename WearWatchFaceContentOverlayEventReported to WearLiveEntryEventReported.
This change renames the atom and its associated enums from WatchFaceContentOverlay to LiveEntry to better reflect the broader scope of the feature, which now includes content from both the Watch Face Overlay and Notifications.
Note - this atom is in development phase, so it is okay to change the name.

Bug: 463714652
Flag: EXEMPT METRICS
Change-Id: I7979fa8b7f6976113930a03cf249e39959b0dea6