Add enum for OTT session duration
This change adds a new event type,
CRITICAL_DURATION_EVENT_TYPE_OTT_SESSION_DURATION_MS, to the
CriticalDurationEventType enum. This will be used to log the total
duration in milliseconds that an OTT call app was active for slicing.

This metric will help in monitoring and understanding the usage of
the OTT slicing feature.

Flag: EXEMPT BUGFIX
Bug: 448567932
Bug: 430509489
Test: m and device test
Change-Id: Iec56b0f045b5e37f4008870d6a6e7053263fc250